separate gdisk call into its own file
