+++ src/gondar.cc
@@ -42,10 +42,7 @@
 #include "device.h"
 #include "shared.h"
 #include "log.h"
-
-// We use gdisk to clean up the GPT such that Windows is happy writing to
-// the disk
-#include "../gdisk/gpt.h"
+#include "gpt_pal.h"
 
 static ssize_t size_t_to_signed(const size_t value) {
   if (value <= SSIZE_MAX) {
@@ -2283,24 +2280,6 @@ DeviceGuyList GetDeviceList() {
   DeviceGuyList device_list;
   GetDevices(&device_list);
   return device_list;
-}
-
-bool clearMbrGpt(char * physical_path) {
-  std::string physical_path_str(physical_path);
-  GPTData gptdata(physical_path_str);
-  // let's get some before/after data on the state of the gpt
-  LOG_INFO << "problems with gpt before reformat=" << gptdata.Verify() << std::endl;
-  gptdata.ClearGPTData();
-  gptdata.MakeProtectiveMBR();
-  int quiet = true;
-  gptdata.SaveGPTData(quiet);
-  int problems = gptdata.Verify();
-  LOG_INFO << "problems with gpt after reformat=" << problems << std::endl;
-  if (problems > 0) {
-    return false;
-  } else {
-    return true;
-  }
 }
 
 bool Install(DeviceGuy* target_device,
